refactor: Divide portfoliodetail and portflois
