+++ src/components/portfoliodetail/DescriptionRow.jsx
