+++ src/components/portfoliodetail/PortfolioCarousel.css
