+++ src/components/portfoliodetail/PortfolioDetailTemplate.jsx
