+++ src/components/portfoliodetail/PriceInfoRow.jsx
